+++ commerce/warehouse/src/main/java/ru/yandex/practicum/warehouse/model/CachedProduct.java
@@ -3,14 +3,16 @@ package ru.yandex.practicum.warehouse.model;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 public class CachedProduct implements ProductInfo {
     private String productId;
     private boolean fragile;
-    private double width;
-    private double height;
-    private double depth;
-    private double weight;
+    private BigDecimal width;
+    private BigDecimal height;
+    private BigDecimal depth;
+    private BigDecimal weight;
     private int quantity;
 }
